+++ src/resources/storage.rs
@@ -3,23 +3,31 @@ use crate::utils::{ChangeTicks, UnsafeUnwrap};
 use atomic_refcell::{AtomicRef, AtomicRefCell, AtomicRefMut};
 use rustc_hash::FxHashMap;
 use std::any::TypeId;
+use std::collections::hash_map::IterMut as HashMapIterMut;
 
+/// Container for a type-erased `Resource` and its `ChangeTicks`.
 pub struct ResourceCell {
-    resource: Box<dyn Resource>,
-    pub ticks: ChangeTicks,
+    value: Box<dyn Resource>,
+    pub(crate) ticks: ChangeTicks,
 }
 
 impl ResourceCell {
-    pub fn new(resource: Box<dyn Resource>, ticks: ChangeTicks) -> Self {
-        Self { resource, ticks }
+    pub(crate) fn new(value: Box<dyn Resource>, ticks: ChangeTicks) -> Self {
+        Self { value, ticks }
     }
 
-    pub fn resource(&self) -> &dyn Resource {
-        &*self.resource
+    /// Returns a type-erased reference to the resource stored inside.
+    pub fn value(&self) -> &dyn Resource {
+        &*self.value
     }
 
-    pub fn resource_mut(&mut self) -> &mut dyn Resource {
-        &mut *self.resource
+    pub(crate) fn value_mut(&mut self) -> &mut dyn Resource {
+        &mut *self.value
+    }
+
+    /// Returns the resource's `ChangeTicks`.
+    pub fn ticks(&self) -> ChangeTicks {
+        self.ticks
     }
 }
 
@@ -38,7 +46,7 @@ impl ResourceStorage {
 
         self.resources
             .insert(TypeId::of::<T>(), AtomicRefCell::new(cell))
-            .map(|c| unsafe { *c.into_inner().resource.downcast().unsafe_unwrap() })
+            .map(|c| unsafe { *c.into_inner().value.downcast().unsafe_unwrap() })
     }
 
     pub fn remove<T>(&mut self) -> Option<T>
@@ -47,7 +55,7 @@ impl ResourceStorage {
     {
         self.resources
             .remove(&TypeId::of::<T>())
-            .map(|c| unsafe { *c.into_inner().resource.downcast().unsafe_unwrap() })
+            .map(|c| unsafe { *c.into_inner().value.downcast().unsafe_unwrap() })
     }
 
     pub fn contains(&self, resource_type_id: &TypeId) -> bool {
@@ -68,5 +76,31 @@ impl ResourceStorage {
         self.resources
             .get(resource_type_id)
             .map(AtomicRefCell::borrow_mut)
+    }
+
+    pub fn iter(&mut self) -> ResourceStorageIter {
+        ResourceStorageIter::new(self)
+    }
+}
+
+pub struct ResourceStorageIter<'a> {
+    inner: HashMapIterMut<'a, TypeId, AtomicRefCell<ResourceCell>>,
+}
+
+impl<'a> ResourceStorageIter<'a> {
+    fn new(resource_storage: &'a mut ResourceStorage) -> Self {
+        Self {
+            inner: resource_storage.resources.iter_mut(),
+        }
+    }
+}
+
+impl<'a> Iterator for ResourceStorageIter<'a> {
+    type Item = (&'a TypeId, &'a ResourceCell);
+
+    fn next(&mut self) -> Option<Self::Item> {
+        self.inner
+            .next()
+            .map(|(type_id, cell)| (type_id, &*cell.get_mut()))
     }
 }